Changes to server backend. Added feature for requesting consolidated data via adress. Fixed bug returning events occuring in the past. Fixed bug in duration and changed return for consolidated to those events with duration < 1 week.

+++ server/server-war/src/java/converter/list/EventRecordConverter.java
@@ -160,7 +160,7 @@ public class EventRecordConverter {
 	 * @return value for duration
 	 */
 	@XmlElement
-	public Short getDuration() {
+	public Long getDuration() {
 		return (expandLevel > 0) ? entity.getDuration() : null;
 	}
 
@@ -170,7 +170,7 @@ public class EventRecordConverter {
 	 * @param value
 	 *            the value to set
 	 */
-	public void setDuration(Short value) {
+	public void setDuration(Long value) {
 		entity.setDuration(value);
 	}
